+++ pkg/workers/textclassifier/textclassifier.go
@@ -125,6 +125,10 @@ func (tc *TextClassifier) processWebArticle(
 	}
 
 	title := strings.TrimSpace(wa.Title)
+	if wa.TranslatedTitle.Valid {
+		title = strings.TrimSpace(wa.TranslatedTitle.String)
+	}
+
 	if len(title) == 0 {
 		logger.Debug().Msg("empty title - web article skipped")
 		return nil