+++ pkg/workers/vectorizer/vectorizer.go
@@ -104,6 +104,10 @@ func (v *Vectorizer) processWebArticle(
 	}
 
 	title := strings.TrimSpace(wa.Title)
+	if wa.TranslatedTitle.Valid {
+		title = strings.TrimSpace(wa.TranslatedTitle.String)
+	}
+
 	if len(title) == 0 {
 		logger.Debug().Msg("empty title - web article skipped")
 		return nil